GH-73991: Add `os.copyfile()` and friends
Move these low-level file copy functions from `shutil` to `os`:

- `copyfileobj()`
- `copyfile()`
- `copymode()`
- `copystat()`

The names remain in `shutil` for backwards compatibility. The higher level
functions (which resemble the `cp` shell utility) also remain in `shutil`.
