+++ Lib/os.py
@@ -23,6 +23,7 @@ and opendir), and leave all pathname manipulation to os.path
 
 #'
 import abc
+import errno
 import sys
 import stat as st
 
@@ -1152,3 +1153,349 @@ if _exists('sched_getaffinity') and sys._get_cpu_count_config() < 0:
 else:
     # Just an alias to cpu_count() (same docstring)
     process_cpu_count = cpu_count
+
+
+# This should never be removed, see rationale in:
+# https://bugs.python.org/issue43743#msg393429
+_USE_CP_SENDFILE = (_exists("sendfile")
+                    and sys.platform.startswith(("linux", "android")))
+_HAS_FCOPYFILE = _exists("_fcopyfile")  # macOS
+COPY_BUFSIZE = 1024 * 1024 if name == 'nt' else 64 * 1024
+
+class CopyError(OSError):
+    pass
+
+class SameFileError(CopyError):
+    """Raised when source and destination are the same file."""
+
+class SpecialFileError(OSError):
+    """Raised when trying to do a kind of operation (e.g. copying) which is
+    not supported on a special file (e.g. a named pipe)"""
+
+class _GiveupOnFastCopy(Exception):
+    """Raised as a signal to fallback on using raw read()/write()
+    file copy when fast-copy functions fail to do so.
+    """
+
+def _fastcopy_fcopyfile(fsrc, fdst, flags):
+    """Copy a regular file content or metadata by using high-performance
+    fcopyfile(3) syscall (macOS).
+    """
+    try:
+        infd = fsrc.fileno()
+        outfd = fdst.fileno()
+    except Exception as err:
+        raise _GiveupOnFastCopy(err)  # not a regular file
+
+    try:
+        _fcopyfile(infd, outfd, flags)
+    except OSError as err:
+        err.filename = fsrc.name
+        err.filename2 = fdst.name
+        if err.errno in {errno.EINVAL, errno.ENOTSUP}:
+            raise _GiveupOnFastCopy(err)
+        else:
+            raise err from None
+
+def _fastcopy_sendfile(fsrc, fdst):
+    """Copy data from one regular mmap-like fd to another by using
+    high-performance sendfile(2) syscall.
+    This should work on Linux >= 2.6.33 only.
+    """
+    # Note: copyfileobj() is left alone in order to not introduce any
+    # unexpected breakage. Possible risks by using zero-copy calls
+    # in copyfileobj() are:
+    # - fdst cannot be open in "a"(ppend) mode
+    # - fsrc and fdst may be open in "t"(ext) mode
+    # - fsrc may be a BufferedReader (which hides unread data in a buffer),
+    #   GzipFile (which decompresses data), HTTPResponse (which decodes
+    #   chunks).
+    # - possibly others (e.g. encrypted fs/partition?)
+    global _USE_CP_SENDFILE
+    try:
+        infd = fsrc.fileno()
+        outfd = fdst.fileno()
+    except Exception as err:
+        raise _GiveupOnFastCopy(err)  # not a regular file
+
+    # Hopefully the whole file will be copied in a single call.
+    # sendfile() is called in a loop 'till EOF is reached (0 return)
+    # so a bufsize smaller or bigger than the actual file size
+    # should not make any difference, also in case the file content
+    # changes while being copied.
+    try:
+        blocksize = max(fstat(infd).st_size, 2 ** 23)  # min 8MiB
+    except OSError:
+        blocksize = 2 ** 27  # 128MiB
+    # On 32-bit architectures truncate to 1GiB to avoid OverflowError,
+    # see bpo-38319.
+    if sys.maxsize < 2 ** 32:
+        blocksize = min(blocksize, 2 ** 30)
+
+    offset = 0
+    while True:
+        try:
+            sent = sendfile(outfd, infd, offset, blocksize)
+        except OSError as err:
+            # ...in oder to have a more informative exception.
+            err.filename = fsrc.name
+            err.filename2 = fdst.name
+
+            if err.errno == errno.ENOTSOCK:
+                # sendfile() on this platform (probably Linux < 2.6.33)
+                # does not support copies between regular files (only
+                # sockets).
+                _USE_CP_SENDFILE = False
+                raise _GiveupOnFastCopy(err)
+
+            if err.errno == errno.ENOSPC:  # filesystem is full
+                raise err from None
+
+            # Give up on first call and if no data was copied.
+            if offset == 0 and lseek(outfd, 0, SEEK_CUR) == 0:
+                raise _GiveupOnFastCopy(err)
+
+            raise err
+        else:
+            if sent == 0:
+                break  # EOF
+            offset += sent
+
+def _fastcopy_readinto(fsrc, fdst, length=COPY_BUFSIZE):
+    """readinto()/memoryview() based variant of copyfileobj().
+    *fsrc* must support readinto() method and both files must be
+    open in binary mode.
+    """
+    # Localize variable access to minimize overhead.
+    fsrc_readinto = fsrc.readinto
+    fdst_write = fdst.write
+    with memoryview(bytearray(length)) as mv:
+        while True:
+            n = fsrc_readinto(mv)
+            if not n:
+                break
+            elif n < length:
+                with mv[:n] as smv:
+                    fdst_write(smv)
+                break
+            else:
+                fdst_write(mv)
+
+def _fastcopy(fsrc, fdst, file_size):
+    # macOS
+    if _HAS_FCOPYFILE:
+        try:
+            _fastcopy_fcopyfile(fsrc, fdst, _COPYFILE_DATA)
+            return
+        except _GiveupOnFastCopy:
+            pass
+    # Linux
+    elif _USE_CP_SENDFILE:
+        try:
+            _fastcopy_sendfile(fsrc, fdst)
+            return
+        except _GiveupOnFastCopy:
+            pass
+    # Windows, see:
+    # https://github.com/python/cpython/pull/7160#discussion_r195405230
+    elif name == 'nt' and file_size > 0:
+        _fastcopy_readinto(fsrc, fdst, min(file_size, COPY_BUFSIZE))
+        return
+
+    copyfileobj(fsrc, fdst)
+
+def copyfileobj(fsrc, fdst, length=0):
+    """copy data from file-like object fsrc to file-like object fdst"""
+    if not length:
+        length = COPY_BUFSIZE
+    # Localize variable access to minimize overhead.
+    fsrc_read = fsrc.read
+    fdst_write = fdst.write
+    while buf := fsrc_read(length):
+        fdst_write(buf)
+
+def _samefile(src, dst):
+    # Macintosh, Unix.
+    if isinstance(src, DirEntry):
+        try:
+            return path.samestat(src.stat(), stat(dst))
+        except OSError:
+            return False
+
+    try:
+        return path.samefile(src, dst)
+    except OSError:
+        return False
+
+def _stat(fn):
+    return fn.stat() if isinstance(fn, DirEntry) else stat(fn)
+
+def _islink(fn):
+    return fn.is_symlink() if isinstance(fn, DirEntry) else path.islink(fn)
+
+def copyfile(src, dst, *, follow_symlinks=True):
+    """Copy data from src to dst in the most efficient way possible.
+
+    If follow_symlinks is not set and src is a symbolic link, a new
+    symlink will be created instead of copying the file it points to.
+
+    """
+    sys.audit("os.copyfile", src, dst)
+
+    if _samefile(src, dst):
+        raise SameFileError("{!r} and {!r} are the same file".format(src, dst))
+
+    file_size = 0
+    for i, fn in enumerate([src, dst]):
+        try:
+            stat_result = _stat(fn)
+        except OSError:
+            # File most likely does not exist
+            pass
+        else:
+            # XXX What about other special files? (sockets, devices...)
+            if st.S_ISFIFO(stat_result.st_mode):
+                fn = fn.path if isinstance(fn, DirEntry) else fn
+                raise SpecialFileError("`%s` is a named pipe" % fn)
+            if name == 'nt' and i == 0:
+                file_size = stat_result.st_size
+
+    if not follow_symlinks and _islink(src):
+        symlink(readlink(src), dst)
+    else:
+        import io
+        with io.open(src, 'rb') as fsrc:
+            try:
+                with io.open(dst, 'wb') as fdst:
+                    _fastcopy(fsrc, fdst, file_size)
+
+            # Issue 43219, raise a less confusing exception
+            except IsADirectoryError as e:
+                if not path.exists(dst):
+                    raise FileNotFoundError(f'Directory does not exist: {dst}') from e
+                else:
+                    raise
+
+    return dst
+
+def copymode(src, dst, *, follow_symlinks=True):
+    """Copy mode bits from src to dst.
+
+    If follow_symlinks is not set, symlinks aren't followed if and only
+    if both `src` and `dst` are symlinks.  If `lchmod` isn't available
+    (e.g. Linux) this method does nothing.
+
+    """
+    sys.audit("os.copymode", src, dst)
+
+    if not follow_symlinks and _islink(src) and path.islink(dst):
+        if _exists('lchmod'):
+            stat_func, chmod_func = lstat, lchmod
+        else:
+            return
+    else:
+        stat_func = _stat
+        if name == 'nt' and path.islink(dst):
+            def chmod_func(*args):
+                chmod(*args, follow_symlinks=True)
+        else:
+            chmod_func = chmod
+
+    stat_result = stat_func(src)
+    chmod_func(dst, st.S_IMODE(stat_result.st_mode))
+
+if _exists('listxattr'):
+    def _copyxattr(src, dst, *, follow_symlinks=True):
+        """Copy extended filesystem attributes from `src` to `dst`.
+
+        Overwrite existing attributes.
+
+        If `follow_symlinks` is false, symlinks won't be followed.
+
+        """
+
+        try:
+            names = listxattr(src, follow_symlinks=follow_symlinks)
+        except OSError as e:
+            if e.errno not in (errno.ENOTSUP, errno.ENODATA, errno.EINVAL):
+                raise
+            return
+        for name in names:
+            try:
+                value = getxattr(src, name, follow_symlinks=follow_symlinks)
+                setxattr(dst, name, value, follow_symlinks=follow_symlinks)
+            except OSError as e:
+                if e.errno not in (errno.EPERM, errno.ENOTSUP, errno.ENODATA,
+                                   errno.EINVAL, errno.EACCES):
+                    raise
+else:
+    def _copyxattr(*args, **kwargs):
+        pass
+
+def copystat(src, dst, *, follow_symlinks=True):
+    """Copy file metadata
+
+    Copy the permission bits, last access time, last modification time, and
+    flags from `src` to `dst`. On Linux, copystat() also copies the "extended
+    attributes" where possible. The file contents, owner, and group are
+    unaffected. `src` and `dst` are path-like objects or path names given as
+    strings.
+
+    If the optional flag `follow_symlinks` is not set, symlinks aren't
+    followed if and only if both `src` and `dst` are symlinks.
+    """
+    sys.audit("os.copystat", src, dst)
+
+    def _nop(*args, ns=None, follow_symlinks=None):
+        pass
+
+    # follow symlinks (aka don't not follow symlinks)
+    follow = follow_symlinks or not (_islink(src) and path.islink(dst))
+    if follow:
+        # use the real function if it exists
+        def lookup(name):
+            return globals().get(name, _nop)
+    else:
+        # use the real function only if it exists
+        # *and* it supports follow_symlinks
+        def lookup(name):
+            fn = globals().get(name, _nop)
+            if fn in supports_follow_symlinks:
+                return fn
+            return _nop
+
+    if isinstance(src, DirEntry):
+        stat_result = src.stat(follow_symlinks=follow)
+    else:
+        stat_result = lookup("stat")(src, follow_symlinks=follow)
+    mode = st.S_IMODE(stat_result.st_mode)
+    lookup("utime")(dst, ns=(stat_result.st_atime_ns, stat_result.st_mtime_ns),
+        follow_symlinks=follow)
+    # We must copy extended attributes before the file is (potentially)
+    # chmod()'ed read-only, otherwise setxattr() will error with -EACCES.
+    _copyxattr(src, dst, follow_symlinks=follow)
+    try:
+        lookup("chmod")(dst, mode, follow_symlinks=follow)
+    except NotImplementedError:
+        # if we got a NotImplementedError, it's because
+        #   * follow_symlinks=False,
+        #   * lchown() is unavailable, and
+        #   * either
+        #       * fchownat() is unavailable or
+        #       * fchownat() doesn't implement AT_SYMLINK_NOFOLLOW.
+        #         (it returned ENOSUP.)
+        # therefore we're out of options--we simply cannot chown the
+        # symlink.  give up, suppress the error.
+        # (which is what shutil always did in this circumstance.)
+        pass
+    if hasattr(st, 'st_flags'):
+        try:
+            lookup("chflags")(dst, stat_result.st_flags, follow_symlinks=follow)
+        except OSError as why:
+            for err in 'EOPNOTSUPP', 'ENOTSUP':
+                if hasattr(errno, err) and why.errno == getattr(errno, err):
+                    break
+            else:
+                raise
+
+__all__.extend(["copyfileobj", "copyfile", "copymode", "copystat"])
