+++ Lib/test/test_shutil.py
@@ -12,16 +12,12 @@ import errno
 import functools
 import pathlib
 import subprocess
-import random
-import string
-import contextlib
-import io
 from shutil import (make_archive,
                     register_archive_format, unregister_archive_format,
                     get_archive_formats, Error, unpack_archive,
                     register_unpack_format, RegistryError,
                     unregister_unpack_format, get_unpack_formats,
-                    SameFileError, _GiveupOnFastCopy)
+                    SameFileError)
 import tarfile
 import zipfile
 try:
@@ -33,6 +29,7 @@ from test import support
 from test.support import os_helper
 from test.support.os_helper import TESTFN, FakePath
 from test.support import warnings_helper
+from test.test_os import read_file, write_file
 
 TESTFN2 = TESTFN + "2"
 TESTFN_SRC = TESTFN + "_SRC"
@@ -70,52 +67,6 @@ def mock_rename(func):
             os.rename = builtin_rename
     return wrap
 
-def write_file(path, content, binary=False):
-    """Write *content* to a file located at *path*.
-
-    If *path* is a tuple instead of a string, os.path.join will be used to
-    make a path.  If *binary* is true, the file will be opened in binary
-    mode.
-    """
-    if isinstance(path, tuple):
-        path = os.path.join(*path)
-    mode = 'wb' if binary else 'w'
-    encoding = None if binary else "utf-8"
-    with open(path, mode, encoding=encoding) as fp:
-        fp.write(content)
-
-def write_test_file(path, size):
-    """Create a test file with an arbitrary size and random text content."""
-    def chunks(total, step):
-        assert total >= step
-        while total > step:
-            yield step
-            total -= step
-        if total:
-            yield total
-
-    bufsize = min(size, 8192)
-    chunk = b"".join([random.choice(string.ascii_letters).encode()
-                      for i in range(bufsize)])
-    with open(path, 'wb') as f:
-        for csize in chunks(size, bufsize):
-            f.write(chunk)
-    assert os.path.getsize(path) == size
-
-def read_file(path, binary=False):
-    """Return contents from a file located at *path*.
-
-    If *path* is a tuple instead of a string, os.path.join will be used to
-    make a path.  If *binary* is true, the file will be opened in binary
-    mode.
-    """
-    if isinstance(path, tuple):
-        path = os.path.join(*path)
-    mode = 'rb' if binary else 'r'
-    encoding = None if binary else "utf-8"
-    with open(path, mode, encoding=encoding) as fp:
-        return fp.read()
-
 def rlistdir(path):
     res = []
     for name in sorted(os.listdir(path)):
@@ -128,36 +79,6 @@ def rlistdir(path):
             res.append(name)
     return res
 
-def supports_file2file_sendfile():
-    # ...apparently Linux and Solaris are the only ones
-    if not hasattr(os, "sendfile"):
-        return False
-    srcname = None
-    dstname = None
-    try:
-        with tempfile.NamedTemporaryFile("wb", dir=os.getcwd(), delete=False) as f:
-            srcname = f.name
-            f.write(b"0123456789")
-
-        with open(srcname, "rb") as src:
-            with tempfile.NamedTemporaryFile("wb", dir=os.getcwd(), delete=False) as dst:
-                dstname = dst.name
-                infd = src.fileno()
-                outfd = dst.fileno()
-                try:
-                    os.sendfile(outfd, infd, 0, 2)
-                except OSError:
-                    return False
-                else:
-                    return True
-    finally:
-        if srcname is not None:
-            os_helper.unlink(srcname)
-        if dstname is not None:
-            os_helper.unlink(dstname)
-
-
-SUPPORTS_SENDFILE = supports_file2file_sendfile()
 
 # AIX 32-bit mode, by default, lacks enough memory for the xz/lzma compiler test
 # The AIX command 'dump -o program' gives XCOFF header information
@@ -2906,399 +2827,6 @@ class TestMove(BaseTest, unittest.TestCase):
             if os.path.exists(TESTFN_DST):
                 os.lchflags(TESTFN_DST, stat.UF_OPAQUE)
                 os_helper.rmtree(TESTFN_DST)
-
-
-class TestCopyFile(unittest.TestCase):
-
-    class Faux(object):
-        _entered = False
-        _exited_with = None
-        _raised = False
-        def __init__(self, raise_in_exit=False, suppress_at_exit=True):
-            self._raise_in_exit = raise_in_exit
-            self._suppress_at_exit = suppress_at_exit
-        def read(self, *args):
-            return ''
-        def __enter__(self):
-            self._entered = True
-        def __exit__(self, exc_type, exc_val, exc_tb):
-            self._exited_with = exc_type, exc_val, exc_tb
-            if self._raise_in_exit:
-                self._raised = True
-                raise OSError("Cannot close")
-            return self._suppress_at_exit
-
-    def test_w_source_open_fails(self):
-        def _open(filename, mode='r'):
-            if filename == 'srcfile':
-                raise OSError('Cannot open "srcfile"')
-            assert 0  # shouldn't reach here.
-
-        with support.swap_attr(shutil, 'open', _open):
-            with self.assertRaises(OSError):
-                shutil.copyfile('srcfile', 'destfile')
-
-    @unittest.skipIf(MACOS, "skipped on macOS")
-    def test_w_dest_open_fails(self):
-        srcfile = self.Faux()
-
-        def _open(filename, mode='r'):
-            if filename == 'srcfile':
-                return srcfile
-            if filename == 'destfile':
-                raise OSError('Cannot open "destfile"')
-            assert 0  # shouldn't reach here.
-
-        with support.swap_attr(shutil, 'open', _open):
-            shutil.copyfile('srcfile', 'destfile')
-        self.assertTrue(srcfile._entered)
-        self.assertTrue(srcfile._exited_with[0] is OSError)
-        self.assertEqual(srcfile._exited_with[1].args,
-                         ('Cannot open "destfile"',))
-
-    @unittest.skipIf(MACOS, "skipped on macOS")
-    def test_w_dest_close_fails(self):
-        srcfile = self.Faux()
-        destfile = self.Faux(True)
-
-        def _open(filename, mode='r'):
-            if filename == 'srcfile':
-                return srcfile
-            if filename == 'destfile':
-                return destfile
-            assert 0  # shouldn't reach here.
-
-        with support.swap_attr(shutil, 'open', _open):
-            shutil.copyfile('srcfile', 'destfile')
-        self.assertTrue(srcfile._entered)
-        self.assertTrue(destfile._entered)
-        self.assertTrue(destfile._raised)
-        self.assertTrue(srcfile._exited_with[0] is OSError)
-        self.assertEqual(srcfile._exited_with[1].args,
-                         ('Cannot close',))
-
-    @unittest.skipIf(MACOS, "skipped on macOS")
-    def test_w_source_close_fails(self):
-
-        srcfile = self.Faux(True)
-        destfile = self.Faux()
-
-        def _open(filename, mode='r'):
-            if filename == 'srcfile':
-                return srcfile
-            if filename == 'destfile':
-                return destfile
-            assert 0  # shouldn't reach here.
-
-        with support.swap_attr(shutil, 'open', _open):
-            with self.assertRaises(OSError):
-                shutil.copyfile('srcfile', 'destfile')
-        self.assertTrue(srcfile._entered)
-        self.assertTrue(destfile._entered)
-        self.assertFalse(destfile._raised)
-        self.assertTrue(srcfile._exited_with[0] is None)
-        self.assertTrue(srcfile._raised)
-
-
-class TestCopyFileObj(unittest.TestCase):
-    FILESIZE = 2 * 1024 * 1024
-
-    @classmethod
-    def setUpClass(cls):
-        write_test_file(TESTFN, cls.FILESIZE)
-
-    @classmethod
-    def tearDownClass(cls):
-        os_helper.unlink(TESTFN)
-        os_helper.unlink(TESTFN2)
-
-    def tearDown(self):
-        os_helper.unlink(TESTFN2)
-
-    @contextlib.contextmanager
-    def get_files(self):
-        with open(TESTFN, "rb") as src:
-            with open(TESTFN2, "wb") as dst:
-                yield (src, dst)
-
-    def assert_files_eq(self, src, dst):
-        with open(src, 'rb') as fsrc:
-            with open(dst, 'rb') as fdst:
-                self.assertEqual(fsrc.read(), fdst.read())
-
-    def test_content(self):
-        with self.get_files() as (src, dst):
-            shutil.copyfileobj(src, dst)
-        self.assert_files_eq(TESTFN, TESTFN2)
-
-    def test_file_not_closed(self):
-        with self.get_files() as (src, dst):
-            shutil.copyfileobj(src, dst)
-            assert not src.closed
-            assert not dst.closed
-
-    def test_file_offset(self):
-        with self.get_files() as (src, dst):
-            shutil.copyfileobj(src, dst)
-            self.assertEqual(src.tell(), self.FILESIZE)
-            self.assertEqual(dst.tell(), self.FILESIZE)
-
-    @unittest.skipIf(os.name != 'nt', "Windows only")
-    def test_win_impl(self):
-        # Make sure alternate Windows implementation is called.
-        with unittest.mock.patch("shutil._copyfileobj_readinto") as m:
-            shutil.copyfile(TESTFN, TESTFN2)
-        assert m.called
-
-        # File size is 2 MiB but max buf size should be 1 MiB.
-        self.assertEqual(m.call_args[0][2], 1 * 1024 * 1024)
-
-        # If file size < 1 MiB memoryview() length must be equal to
-        # the actual file size.
-        with tempfile.NamedTemporaryFile(dir=os.getcwd(), delete=False) as f:
-            f.write(b'foo')
-        fname = f.name
-        self.addCleanup(os_helper.unlink, fname)
-        with unittest.mock.patch("shutil._copyfileobj_readinto") as m:
-            shutil.copyfile(fname, TESTFN2)
-        self.assertEqual(m.call_args[0][2], 3)
-
-        # Empty files should not rely on readinto() variant.
-        with tempfile.NamedTemporaryFile(dir=os.getcwd(), delete=False) as f:
-            pass
-        fname = f.name
-        self.addCleanup(os_helper.unlink, fname)
-        with unittest.mock.patch("shutil._copyfileobj_readinto") as m:
-            shutil.copyfile(fname, TESTFN2)
-        assert not m.called
-        self.assert_files_eq(fname, TESTFN2)
-
-
-class _ZeroCopyFileTest(object):
-    """Tests common to all zero-copy APIs."""
-    FILESIZE = (10 * 1024 * 1024)  # 10 MiB
-    FILEDATA = b""
-    PATCHPOINT = ""
-
-    @classmethod
-    def setUpClass(cls):
-        write_test_file(TESTFN, cls.FILESIZE)
-        with open(TESTFN, 'rb') as f:
-            cls.FILEDATA = f.read()
-            assert len(cls.FILEDATA) == cls.FILESIZE
-
-    @classmethod
-    def tearDownClass(cls):
-        os_helper.unlink(TESTFN)
-
-    def tearDown(self):
-        os_helper.unlink(TESTFN2)
-
-    @contextlib.contextmanager
-    def get_files(self):
-        with open(TESTFN, "rb") as src:
-            with open(TESTFN2, "wb") as dst:
-                yield (src, dst)
-
-    def zerocopy_fun(self, *args, **kwargs):
-        raise NotImplementedError("must be implemented in subclass")
-
-    def reset(self):
-        self.tearDown()
-        self.tearDownClass()
-        self.setUpClass()
-        self.setUp()
-
-    # ---
-
-    def test_regular_copy(self):
-        with self.get_files() as (src, dst):
-            self.zerocopy_fun(src, dst)
-        self.assertEqual(read_file(TESTFN2, binary=True), self.FILEDATA)
-        # Make sure the fallback function is not called.
-        with self.get_files() as (src, dst):
-            with unittest.mock.patch('shutil.copyfileobj') as m:
-                shutil.copyfile(TESTFN, TESTFN2)
-            assert not m.called
-
-    def test_same_file(self):
-        self.addCleanup(self.reset)
-        with self.get_files() as (src, dst):
-            with self.assertRaises((OSError, _GiveupOnFastCopy)):
-                self.zerocopy_fun(src, src)
-        # Make sure src file is not corrupted.
-        self.assertEqual(read_file(TESTFN, binary=True), self.FILEDATA)
-
-    def test_non_existent_src(self):
-        name = tempfile.mktemp(dir=os.getcwd())
-        with self.assertRaises(FileNotFoundError) as cm:
-            shutil.copyfile(name, "new")
-        self.assertEqual(cm.exception.filename, name)
-
-    def test_empty_file(self):
-        srcname = TESTFN + 'src'
-        dstname = TESTFN + 'dst'
-        self.addCleanup(lambda: os_helper.unlink(srcname))
-        self.addCleanup(lambda: os_helper.unlink(dstname))
-        with open(srcname, "wb"):
-            pass
-
-        with open(srcname, "rb") as src:
-            with open(dstname, "wb") as dst:
-                self.zerocopy_fun(src, dst)
-
-        self.assertEqual(read_file(dstname, binary=True), b"")
-
-    def test_unhandled_exception(self):
-        with unittest.mock.patch(self.PATCHPOINT,
-                                 side_effect=ZeroDivisionError):
-            self.assertRaises(ZeroDivisionError,
-                              shutil.copyfile, TESTFN, TESTFN2)
-
-    def test_exception_on_first_call(self):
-        # Emulate a case where the first call to the zero-copy
-        # function raises an exception in which case the function is
-        # supposed to give up immediately.
-        with unittest.mock.patch(self.PATCHPOINT,
-                                 side_effect=OSError(errno.EINVAL, "yo")):
-            with self.get_files() as (src, dst):
-                with self.assertRaises(_GiveupOnFastCopy):
-                    self.zerocopy_fun(src, dst)
-
-    def test_filesystem_full(self):
-        # Emulate a case where filesystem is full and sendfile() fails
-        # on first call.
-        with unittest.mock.patch(self.PATCHPOINT,
-                                 side_effect=OSError(errno.ENOSPC, "yo")):
-            with self.get_files() as (src, dst):
-                self.assertRaises(OSError, self.zerocopy_fun, src, dst)
-
-
-@unittest.skipIf(not SUPPORTS_SENDFILE, 'os.sendfile() not supported')
-class TestZeroCopySendfile(_ZeroCopyFileTest, unittest.TestCase):
-    PATCHPOINT = "os.sendfile"
-
-    def zerocopy_fun(self, fsrc, fdst):
-        return shutil._fastcopy_sendfile(fsrc, fdst)
-
-    def test_non_regular_file_src(self):
-        with io.BytesIO(self.FILEDATA) as src:
-            with open(TESTFN2, "wb") as dst:
-                with self.assertRaises(_GiveupOnFastCopy):
-                    self.zerocopy_fun(src, dst)
-                shutil.copyfileobj(src, dst)
-
-        self.assertEqual(read_file(TESTFN2, binary=True), self.FILEDATA)
-
-    def test_non_regular_file_dst(self):
-        with open(TESTFN, "rb") as src:
-            with io.BytesIO() as dst:
-                with self.assertRaises(_GiveupOnFastCopy):
-                    self.zerocopy_fun(src, dst)
-                shutil.copyfileobj(src, dst)
-                dst.seek(0)
-                self.assertEqual(dst.read(), self.FILEDATA)
-
-    def test_exception_on_second_call(self):
-        def sendfile(*args, **kwargs):
-            if not flag:
-                flag.append(None)
-                return orig_sendfile(*args, **kwargs)
-            else:
-                raise OSError(errno.EBADF, "yo")
-
-        flag = []
-        orig_sendfile = os.sendfile
-        with unittest.mock.patch('os.sendfile', create=True,
-                                 side_effect=sendfile):
-            with self.get_files() as (src, dst):
-                with self.assertRaises(OSError) as cm:
-                    shutil._fastcopy_sendfile(src, dst)
-        assert flag
-        self.assertEqual(cm.exception.errno, errno.EBADF)
-
-    def test_cant_get_size(self):
-        # Emulate a case where src file size cannot be determined.
-        # Internally bufsize will be set to a small value and
-        # sendfile() will be called repeatedly.
-        with unittest.mock.patch('os.fstat', side_effect=OSError) as m:
-            with self.get_files() as (src, dst):
-                shutil._fastcopy_sendfile(src, dst)
-                assert m.called
-        self.assertEqual(read_file(TESTFN2, binary=True), self.FILEDATA)
-
-    def test_small_chunks(self):
-        # Force internal file size detection to be smaller than the
-        # actual file size. We want to force sendfile() to be called
-        # multiple times, also in order to emulate a src fd which gets
-        # bigger while it is being copied.
-        mock = unittest.mock.Mock()
-        mock.st_size = 65536 + 1
-        with unittest.mock.patch('os.fstat', return_value=mock) as m:
-            with self.get_files() as (src, dst):
-                shutil._fastcopy_sendfile(src, dst)
-                assert m.called
-        self.assertEqual(read_file(TESTFN2, binary=True), self.FILEDATA)
-
-    def test_big_chunk(self):
-        # Force internal file size detection to be +100MB bigger than
-        # the actual file size. Make sure sendfile() does not rely on
-        # file size value except for (maybe) a better throughput /
-        # performance.
-        mock = unittest.mock.Mock()
-        mock.st_size = self.FILESIZE + (100 * 1024 * 1024)
-        with unittest.mock.patch('os.fstat', return_value=mock) as m:
-            with self.get_files() as (src, dst):
-                shutil._fastcopy_sendfile(src, dst)
-                assert m.called
-        self.assertEqual(read_file(TESTFN2, binary=True), self.FILEDATA)
-
-    def test_blocksize_arg(self):
-        with unittest.mock.patch('os.sendfile',
-                                 side_effect=ZeroDivisionError) as m:
-            self.assertRaises(ZeroDivisionError,
-                              shutil.copyfile, TESTFN, TESTFN2)
-            blocksize = m.call_args[0][3]
-            # Make sure file size and the block size arg passed to
-            # sendfile() are the same.
-            self.assertEqual(blocksize, os.path.getsize(TESTFN))
-            # ...unless we're dealing with a small file.
-            os_helper.unlink(TESTFN2)
-            write_file(TESTFN2, b"hello", binary=True)
-            self.addCleanup(os_helper.unlink, TESTFN2 + '3')
-            self.assertRaises(ZeroDivisionError,
-                              shutil.copyfile, TESTFN2, TESTFN2 + '3')
-            blocksize = m.call_args[0][3]
-            self.assertEqual(blocksize, 2 ** 23)
-
-    def test_file2file_not_supported(self):
-        # Emulate a case where sendfile() only support file->socket
-        # fds. In such a case copyfile() is supposed to skip the
-        # fast-copy attempt from then on.
-        assert shutil._USE_CP_SENDFILE
-        try:
-            with unittest.mock.patch(
-                    self.PATCHPOINT,
-                    side_effect=OSError(errno.ENOTSOCK, "yo")) as m:
-                with self.get_files() as (src, dst):
-                    with self.assertRaises(_GiveupOnFastCopy):
-                        shutil._fastcopy_sendfile(src, dst)
-                assert m.called
-            assert not shutil._USE_CP_SENDFILE
-
-            with unittest.mock.patch(self.PATCHPOINT) as m:
-                shutil.copyfile(TESTFN, TESTFN2)
-                assert not m.called
-        finally:
-            shutil._USE_CP_SENDFILE = True
-
-
-@unittest.skipIf(not MACOS, 'macOS only')
-class TestZeroCopyMACOS(_ZeroCopyFileTest, unittest.TestCase):
-    PATCHPOINT = "posix._fcopyfile"
-
-    def zerocopy_fun(self, src, dst):
-        return shutil._fastcopy_fcopyfile(src, dst, posix._COPYFILE_DATA)
 
 
 class TestGetTerminalSize(unittest.TestCase):
